[PATCH] Fix connection state checking in NET_SendToAll()
Fix NET_SendToAll() so that it really skips non-connected clients. Rename the
cryptic state1 and state2 variables to something more indicative of their
purpose. While we're at it, change the type of the blocktime argument to
double (the comparison is against a double.)

Based on patch by O.Sezer <sezero@users.sourceforge.net>

+++ NQ/net.h
@@ -293,7 +293,7 @@ int NET_SendUnreliableMessage(struct qsocket_s *sock, sizebuf_t *data);
 /*
  * This is a reliable *blocking* send to all attached clients.
  */
-int NET_SendToAll(sizebuf_t *data, int blocktime);
+int NET_SendToAll(sizebuf_t *data, double blocktime);
 
 /*
  * if a dead connection is returned by a get or send function, this function
